Keep the protected exit in the footer, and retire the fixtures the modal took with it (#388)
Adding toggle, reorder, and unbind hints to the Settings footer pushed the
lowest-priority hint off the end of the bar — and that hint was Ctrl-Q. A
footer that lists every control and omits the way out is the wrong trade, so
it lists the controls that fit and always the exit; the whole set lives in
Help and in the Keys section itself.

The three keys-editor ledger fixtures and the mouse-action fixture drove the
standalone Keys modal, waiting on a title no longer rendered anywhere. They
retire with it: the twenty-one Settings scenarios cover validation, dirty,
unbind, reset, recovery, and reduced geometry through the editor that
replaced it, and the confirm-button mouse target keeps its unit coverage.

The Settings reducer is split in two along the seam that was already there:
the draft's own authority, and what the registry editors ask of it.
