+++ dev-docs/tmux-scenarios/settings-agent-types-error.json
@@ -164,7 +164,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
