+++ dev-docs/tmux-scenarios/settings-agent-types-focused.json
@@ -98,7 +98,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
