+++ dev-docs/tmux-scenarios/settings-agent-types-normal.json
@@ -85,7 +85,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
