+++ dev-docs/tmux-scenarios/settings-agent-types-recovery.json
@@ -109,7 +109,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
