+++ dev-docs/tmux-scenarios/settings-keys-focused.json
@@ -108,7 +108,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
