+++ dev-docs/tmux-scenarios/settings-keys-normal.json
@@ -94,7 +94,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
