+++ dev-docs/tmux-scenarios/settings-keys-unavailable.json
@@ -98,7 +98,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
