+++ dev-docs/tmux-scenarios/settings-screens-focused.json
@@ -103,7 +103,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
