+++ dev-docs/tmux-scenarios/settings-screens-normal.json
@@ -90,7 +90,7 @@
     {
       "op": "assert-frame",
       "contains": [
-        "Tab pane"
+        "quit"
       ],
       "absent": []
     },
