--- dev-docs/tmux-scenarios/v1/keys-editor-recovery-small.json
@@ -1,76 +1,0 @@
-{
-  "schema": 1,
-  "name": "keys-editor-recovery-small",
-  "platform": "macos",
-  "terminal": { "cols": 120, "rows": 36 },
-  "workspace": {
-    "mode": 448,
-    "dirs": [
-      { "path": "config", "mode": 448 },
-      { "path": "work", "mode": 448 }
-    ],
-    "files": [
-      {
-        "path": "config/settings.toml",
-        "content": { "utf8": "# retain malformed keymap\nsettings_schema = 2\n[keymap.global]\n\"core.open-keys\" = [\"Ctrl+\"]\n[extensions.future]\nvalue = \"stay\"\n" },
-        "mode": 384
-      }
-    ],
-    "env": []
-  },
-  "steps": [
-    {
-      "op": "launch",
-      "argv": ["jefe", "--config", "${workspace}/config"],
-      "env": [],
-      "cwd": "work"
-    },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "LLxprt Jefe",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "f9", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "KEY-E401",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["Keys - Keyboard Bindings", "Recovery", "Save disabled"],
-      "absent": []
-    },
-    { "op": "resize", "size": { "cols": 44, "rows": 10 } },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Keys (compact)",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["Esc Back", "Ctrl-Q Quit"],
-      "absent": []
-    },
-    { "op": "key", "key": "escape", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Agent Types",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-file",
-      "file": {
-        "path": "config/settings.toml",
-        "content": { "utf8": "# retain malformed keymap\nsettings_schema = 2\n[keymap.global]\n\"core.open-keys\" = [\"Ctrl+\"]\n[extensions.future]\nvalue = \"stay\"\n" }
-      }
-    },
-    { "op": "key", "key": "q", "modifiers": ["control"] },
-    { "op": "finish" }
-  ],
-  "secrets": []
-}
