--- dev-docs/tmux-scenarios/v1/keys-editor.json
@@ -1,156 +1,0 @@
-{
-  "schema": 1,
-  "name": "keys-editor",
-  "platform": "macos",
-  "terminal": { "cols": 120, "rows": 36 },
-  "workspace": {
-    "mode": 448,
-    "dirs": [
-      { "path": "config", "mode": 448 },
-      { "path": "work", "mode": 448 }
-    ],
-    "files": [
-      {
-        "path": "config/settings.toml",
-        "content": { "utf8": "# keep\nsettings_schema = 2\n[extensions.future]\nvalue = 'stay'\n" },
-        "mode": 384
-      }
-    ],
-    "env": []
-  },
-  "steps": [
-    {
-      "op": "launch",
-      "argv": ["jefe", "--config", "${workspace}/config"],
-      "env": [],
-      "cwd": "work"
-    },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "LLxprt Jefe",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "f9", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Keys - Keyboard Bindings",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["core.emergency-exit", "core.open-keys", "Esc Back", "Ctrl-Q Quit"],
-      "absent": ["KEY-E401"]
-    },
-    { "op": "key", "key": "down", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "> core.open-keys",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "e", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Editing core.open-keys",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["Editing core.open-keys", "> core.open-keys"],
-      "absent": []
-    },
-    { "op": "key", "key": "backspace", "modifiers": [] },
-    { "op": "key", "key": "backspace", "modifiers": [] },
-    { "op": "key", "key": "f", "modifiers": ["shift"] },
-    { "op": "key", "key": "1", "modifiers": [] },
-    { "op": "key", "key": "2", "modifiers": [] },
-    { "op": "key", "key": "enter", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "KEY-E401",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["KEY-E401", "Save disabled"],
-      "absent": ["Save enabled"]
-    },
-    { "op": "key", "key": "backspace", "modifiers": [] },
-    { "op": "key", "key": "backspace", "modifiers": [] },
-    { "op": "key", "key": "backspace", "modifiers": [] },
-    { "op": "key", "key": ".", "modifiers": [] },
-    { "op": "key", "key": "enter", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Save enabled",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["Unsaved changes", "Save enabled", "."],
-      "absent": ["KEY-E401"]
-    },
-    { "op": "key", "key": "escape", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Save / Discard / Cancel",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "escape", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "actions.run-down",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "down", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "> actions.dispatch",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "up", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "> core.open-keys",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-frame",
-      "contains": ["Keys - Keyboard Bindings", "Unsaved changes"],
-      "absent": ["Save / Discard / Cancel"]
-    },
-    { "op": "key", "key": "s", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "LLxprt Jefe",
-      "timeout_ms": 15000
-    },
-    {
-      "op": "assert-file",
-      "file": {
-        "path": "config/settings.toml",
-        "content": { "utf8": "# keep\nsettings_schema = 2\n[extensions.future]\nvalue = 'stay'\n[keymap.\"global\"]\n\"core.open-keys\" = [\".\"]\n" }
-      }
-    },
-    { "op": "key", "key": ".", "modifiers": [] },
-    {
-      "op": "wait",
-      "source": "frame",
-      "literal": "Keys - Keyboard Bindings",
-      "timeout_ms": 15000
-    },
-    { "op": "key", "key": "q", "modifiers": ["control"] },
-    { "op": "finish" }
-  ],
-  "secrets": []
-}
