--- dev-docs/tmux-scenarios/v1/mouse-action-consistency.json
@@ -1,54 +1,0 @@
-{
-  "schema": 1,
-  "name": "mouse-action-consistency",
-  "platform": "macos",
-  "terminal": { "cols": 120, "rows": 36 },
-  "workspace": {
-    "mode": 448,
-    "dirs": [
-      { "path": "config", "mode": 448 },
-      { "path": "work", "mode": 448 }
-    ],
-    "files": [
-      {
-        "path": "config/settings.toml",
-        "content": { "utf8": "settings_schema = 2\n" },
-        "mode": 384
-      },
-      {
-        "path": "config/state.json",
-        "content": {
-          "utf8": "{\n  \"schema_version\": 1,\n  \"repositories\": [{\n    \"id\": \"repo\",\n    \"name\": \"Mouse Repo\",\n    \"slug\": \"mouse-repo\",\n    \"base_dir\": \"/tmp\",\n    \"default_profile\": \"\",\n    \"remote\": { \"enabled\": false, \"login_user\": \"\", \"host\": \"\", \"port\": 22, \"run_as_user\": \"\" },\n    \"agent_ids\": []\n  }],\n  \"agents\": [],\n  \"selected_repository_index\": 0,\n  \"selected_agent_index\": null\n}\n"
-        },
-        "mode": 384
-      }
-    ],
-    "env": []
-  },
-  "steps": [
-    {
-      "op": "launch",
-      "argv": ["jefe", "--config", "${workspace}/config"],
-      "env": [],
-      "cwd": "work"
-    },
-    { "op": "wait", "source": "frame", "literal": "Mouse Repo", "timeout_ms": 15000 },
-    { "op": "key", "key": "f9", "modifiers": [] },
-    { "op": "wait", "source": "frame", "literal": "Keys - Keyboard Bindings", "timeout_ms": 15000 },
-    { "op": "key", "key": "down", "modifiers": [] },
-    { "op": "key", "key": "u", "modifiers": [] },
-    { "op": "wait", "source": "frame", "literal": "Unsaved changes", "timeout_ms": 15000 },
-    { "op": "text", "text": "\u001b[<0;4;7M\u001b[<32;5;7M\u001b[<0;5;7m" },
-    { "op": "assert-frame", "contains": ["Unsaved changes", "core.open-keys"], "absent": [] },
-    { "op": "text", "text": "\u001b[<0;119;35M\u001b[<0;119;35m" },
-    { "op": "assert-frame", "contains": ["Unsaved changes"], "absent": [] },
-    { "op": "key", "key": "escape", "modifiers": [] },
-    { "op": "wait", "source": "frame", "literal": "Save / Discard / Cancel", "timeout_ms": 15000 },
-    { "op": "key", "key": "left", "modifiers": [] },
-    { "op": "key", "key": "enter", "modifiers": [] },
-    { "op": "wait", "source": "frame", "literal": "Mouse Repo", "timeout_ms": 15000 },
-    { "op": "key", "key": "q", "modifiers": ["control"] },
-    { "op": "finish" }
-  ],
-  "secrets": []
-}
